MFRC522: improve RPi.GPIO usage
Instead of letting the users of an MFRC522 instance clean up the resources
of RPi.GPIO, do this in the __del__ method of the instance. Also only clean
up the channel used as a complete cleanup might interfere with other parts
of a program that uses RPi.GPIO also for other stuff.